requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0

# LLM / VLM
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Diffusion
diffusers>=0.25.0
controlnet-aux>=0.0.7
invisible-watermark>=0.2.0

# Video
opencv-python>=4.8.0
decord>=0.6.0
ffmpeg-python>=0.2.0

# Scene / depth (optional: install separately if needed)
# timm for DPT/MiDaS

# Config & validation
pydantic>=2.0.0
PyYAML>=6.0

# CLI & logging
tqdm>=4.66.0

# Optional: detection for annotations
# ultralytics>=8.0.0

# Utilities
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0

# Tests
pytest>=7.0.0
